Upload timestamp (#129)
* feat: validate request signature with timestamp

* Uploads with no parcels are marked as unathorized

When asking permissions for uploading content for a land, it would
check that the address was the owner of all the parcels. If no
parcel was provided, it would authorize the request (though it'd
be invalidated when size > 0). This was an oddly and error prone
validation.

* timestamp was interpreted as string from redis

In the metadata struct, timestamp is an int and so should be
for the redis layer

* fix integration tests

* move private key to environment