remove get to marketplace on /mappings
There was an issue when request /mappings where the content server
would return information about indesired parcels.
The issue was that the /mappings request was reading all parcels of
the involved states of the rect requested. And ultimately, the
request to the decentraland api was completely unnecessary

We completely remove the requests to the api except for the authorized
request on uploads